replace Section with NewSection
