+++ src/routes/trading-view/top-list/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import Breadcrumbs from '$lib/breadcrumb/Breadcrumbs.svelte';
-	import { ContentCard, ContentCardsSection, HeroBanner, NewSection } from '$lib/components';
+	import { ContentCard, ContentCardsSection, HeroBanner, Section } from '$lib/components';
 </script>
 
 <svelte:head>
@@ -11,9 +11,9 @@
 <Breadcrumbs labels={{ 'top-list': 'Top lists' }} />
 
 <main>
-	<NewSection tag="header">
+	<Section tag="header">
 		<HeroBanner contentFullWidth title="Top lists" subtitle="Explore currently the most volatile trading pairs." />
-	</NewSection>
+	</Section>
 
 	<ContentCardsSection>
 		<ContentCard
