+++ src/routes/trading-view/top-list/[direction=momentum]/+page.svelte
@@ -5,7 +5,7 @@
 	import type { PageData } from './$types';
 	import Breadcrumbs from '$lib/breadcrumb/Breadcrumbs.svelte';
 	import MomentumTable from '$lib/momentum/MomentumTable.svelte';
-	import { HeroBanner, NewSection } from '$lib/components';
+	import { HeroBanner, Section } from '$lib/components';
 
 	export let data: PageData;
 	$: up = data.direction === 'up';
@@ -19,7 +19,7 @@
 <Breadcrumbs labels={{ 'top-list': 'Top lists', 'daily-up': 'Daily gainers', 'daily-down': 'Daily losers' }} />
 
 <main>
-	<NewSection tag="header">
+	<Section tag="header">
 		<HeroBanner contentFullWidth title="Trading pairs with the most {up ? 'profit' : 'loss'} for the last 24h">
 			<svelte:fragment slot="subtitle">
 				<a class="body-link" href="/trading-view/trading-pairs">Trading pairs</a>
@@ -28,11 +28,11 @@
 				today. Showing only the pairs with minimum $1M liquidity. All trading pairs are benchmarked against the US Dollar.
 			</svelte:fragment>
 		</HeroBanner>
-	</NewSection>
+	</Section>
 
-	<NewSection>
+	<Section>
 		<MomentumTable pairs={data.pairs} />
-	</NewSection>
+	</Section>
 </main>
 
 <style>
